Separate the logging configuration from the setup
The configuration can still be loaded in its previous form by requiring
'viewpoint/logging/config'.

Resolves #121
