+++ spec/spec_helper.rb
@@ -1,5 +1,6 @@
 $: << File.dirname(__FILE__) + '/../lib/'
 require 'viewpoint'
+require 'viewpoint/logging/config'
 require 'ostruct'
 require 'turn/autorun'
 require_relative 'xml_matcher'
